update routes, move components and views to appropriate folder
