--- client/src/assets/components/DetectFormData.tsx
@@ -1,10 +1,0 @@
-function DetectFormData(formData: object) {
-  for (const key in formData) {
-    if ((formData as any)[key] !== '') {
-      return true;
-    }
-  }
-  return false;
-}
-
-export default DetectFormData;
