--- client/src/assets/components/DisplayGrids/index.ts
@@ -1,3 +1,0 @@
-export { default as BulletGrid } from './BulletGrid';
-export { default as InstructionGrid } from './InstructionGrid';
-export { default as GridImages } from './GridImages';
