--- client/src/assets/components/Forms/index.ts
@@ -1,5 +1,0 @@
-export { default as FileUploadInput } from './FileUploadInput';
-export { default as LinkPage } from './LinkPage';
-export { default as RadioGroup } from './RadioGroup';
-export { default as Select } from './Select';
-export { default as TextField } from './TextField';
