--- client/src/features/action/ActionForm.tsx
@@ -1,19 +1,0 @@
-import { LinkPage } from '../../assets/components/Forms';
-import { TextLink } from '../../types';
-
-const links: TextLink[] = [
-  {
-    url: '/need/',
-    text: 'Share a Need',
-  },
-  {
-    url: '/offer/',
-    text: 'Make an Offer',
-  },
-];
-
-function ActionForm() {
-  return <LinkPage title="What would you like to do?" links={links} />;
-}
-
-export default ActionForm;
