--- client/src/features/action/offers/OfferFormGoods.tsx
@@ -1,170 +1,0 @@
-import * as React from 'react';
-import { Grid } from '@mui/material';
-import { Button } from '@mui/material';
-
-import { FileUploadInput, RadioGroup, Select, TextField } from '../../../assets/components/Forms';
-import NeedOfferForm from '../NeedOfferForm';
-
-const categories = [
-  { value: 'figs', text: 'Figs' },
-  { value: 'peaches', text: 'Peaches' },
-  { value: 'pears', text: 'Pears' },
-];
-const conditions = [
-  { value: 'like-new', text: 'Like new' },
-  { value: 'excellent', text: 'Excellent' },
-  { value: 'good', text: 'Good' },
-];
-const offerTypes = [
-  { value: 'donation', text: 'Donation' },
-  { value: 'short-term', text: 'Short term loan (<1 month)' },
-  { value: 'long-term', text: 'Long term loan (>1 month)' },
-];
-const deliveryTypes = [
-  { value: 'pick-up', text: 'Pick up only' },
-  { value: 'drop-off', text: 'Drop off only' },
-  { value: 'pick-up-drop-off', text: 'Pick up or drop off' },
-];
-
-interface FormData {
-  title: string;
-  location: string;
-  description: string;
-  category: string;
-  condition: string;
-  quantity: string;
-  offerType: string;
-  deliveryMethod: string;
-}
-
-const initialFormData: FormData = {
-  title: '',
-  location: '',
-  description: '',
-  category: '',
-  condition: '',
-  quantity: '',
-  offerType: '',
-  deliveryMethod: '',
-};
-
-function OfferFormGoods() {
-  const [formData, setFormData] = React.useState<FormData>(initialFormData);
-
-  // HTMLInputElement does not work for the MUISelect - This works, but can we find a better way of doing it?
-  const handleChange = (
-    event:
-      | React.ChangeEvent<HTMLInputElement>
-      | React.ChangeEvent<{ name?: string | undefined; value: unknown }>,
-  ): void => {
-    let { name = '', value }: { name?: string | undefined; value: unknown } = event.target;
-    setFormData((fData) => ({
-      ...fData,
-      [name]: value,
-    }));
-  };
-
-  return (
-    <NeedOfferForm title="Make an Offer: Goods">
-      <Grid container spacing={5}>
-        <Grid item md={8} xs={12}>
-          <TextField
-            id="title"
-            label="Title"
-            placeholder="What goods do you need?"
-            value={formData.title}
-            onChange={handleChange}
-          />
-        </Grid>
-        <Grid item md={4} xs={12}>
-          <TextField
-            id="location"
-            label="Location"
-            placeholder="City, State"
-            value={formData.location}
-            onChange={handleChange}
-          />
-        </Grid>
-        <Grid item md={12} xs={12}>
-          <TextField
-            id="description"
-            label="Description"
-            placeholder="Describe what goods you are looking for"
-            value={formData.description}
-            onChange={handleChange}
-            isMultiline={true}
-          />
-        </Grid>
-        <Grid item md={8} xs={12}>
-          <Select
-            id="category"
-            label="Category"
-            placeholder="Select a category"
-            options={categories}
-            value={formData.category}
-            onChange={handleChange}
-          />
-        </Grid>
-        <Grid item md={8} xs={12}>
-          <Select
-            id="condition"
-            label="Condition"
-            placeholder="Select a preferred condition"
-            options={conditions}
-            value={formData.condition}
-            onChange={handleChange}
-          />
-        </Grid>
-        <Grid item md={4} xs={12}>
-          <TextField
-            id="quantity"
-            label="Quantity"
-            placeholder="# of goods needed"
-            value={formData.quantity}
-            onChange={handleChange}
-          />
-        </Grid>
-        <Grid item md={8} xs={12}>
-          <RadioGroup
-            label="Offer Type"
-            id="offerType"
-            options={offerTypes}
-            value={formData.offerType}
-            onChange={handleChange}
-          />
-        </Grid>
-        <Grid item md={4} xs={12}>
-          <RadioGroup
-            label="Delivery Method"
-            id="deliveryMethod"
-            options={deliveryTypes}
-            value={formData.deliveryMethod}
-            onChange={handleChange}
-          />
-        </Grid>
-        <Grid item xs={12}>
-          <FileUploadInput
-            label="Photos"
-            id="photos"
-            text="Click here to upload photos"
-            onChange={handleChange}
-          />
-        </Grid>
-        <Grid item container xs={12} justifyContent="center">
-          <Grid item>
-            <Button variant="contained" color="secondary">
-              Save Draft
-            </Button>
-          </Grid>
-          <Grid item>
-            <Button variant="contained" color="primary">
-              Submit Offer
-            </Button>
-          </Grid>
-        </Grid>
-      </Grid>
-    </NeedOfferForm>
-  );
-}
-
-export default OfferFormGoods;
