--- client/src/features/static/about/OurStory.tsx
@@ -1,7 +1,0 @@
-import * as React from 'react';
-
-function OurStory() {
-  return <div>OurStory</div>;
-}
-
-export default OurStory;
