--- client/src/features/static/policy/PrivacyPolicy.tsx
@@ -1,7 +1,0 @@
-import * as React from 'react';
-
-function PrivacyPolicy() {
-  return <div>PrivacyPolicy</div>;
-}
-
-export default PrivacyPolicy;
