--- client/src/features/static/policy/TermsOfService.tsx
@@ -1,7 +1,0 @@
-import * as React from 'react';
-
-function TermsOfService() {
-  return <div>TermsOfService</div>;
-}
-
-export default TermsOfService;
