--- client/src/features/support/ContactUs.tsx
@@ -1,7 +1,0 @@
-import * as React from 'react';
-
-function ContactUs() {
-  return <div>ContactUs</div>;
-}
-
-export default ContactUs;
