--- client/src/features/support/Help.tsx
@@ -1,7 +1,0 @@
-import * as React from 'react';
-
-function Help() {
-  return <div>Help</div>;
-}
-
-export default Help;
