--- client/src/features/users/User.tsx
@@ -1,7 +1,0 @@
-import * as React from 'react';
-
-function User() {
-  return <div>User Profile</div>;
-}
-
-export default User;
